Split scratchstack-database ops/iam/role up
